Add Lab 1: Position in sequence
A iterative function, that searches for a given element in a sequence of signed integers. The sequence is either monotonically increasing or decreasing. The function takes the sequence and the target element as input and return the index of the found element in the sequence. If the element is not found, the function returns -1.