Add Lab 4: Heap Q
A heap that acts as a priority queue for the instances of the person structure.
The program supports following operations:

+ n v_1 ... v_n - add n individuals (format: <date> <name>) to the priority queue (heap). Initially, the first name in alphabetical order is on the top;
- m - remove m elements from the top of the heap and print their data on the screen (format: <date> <name> e.g. 19-04-2022 Alan);
p - print all elements of the heap in the order in which they are stored in an array;
r - rearrange the order of the queue such that the earliest date is on top. Repeats of this operation alternate between the two orders;
q - exit the program (and deallocate memory).