Add Lab 5: Hash table
Counts the number of occurences of each word in the text given in the input. Store these numbers in a hash table: the word is interpreted as a key (type string), and the number of occurences as a value (type int).

The text (any number of words with only lower case letters, separated by spaces or new line characters). The words starting with character '?' and '-', and the word "q" are interpreted as commands:

 '?': print the number of occurences of the given word on the screen,
 '-': remove the information about the given word from your hash table (from this point you should start counting the occurences of this word from 0),
 'q': exit.